@@ -5,22 +5,26 @@ import Layout from './components/Layout'
 import GuildPicker from './components/GuildPicker'
 import GuildCompare from './components/GuildCompare'
 import SkillPicker from './components/SkillPicker'
-import { MuiThemeProvider, createTheme } from '@material-ui/core/styles';
+import { ThemeProvider, StyledEngineProvider, createTheme, adaptV4Theme } from '@mui/material/styles';
 import themeOverrides from './theme.js';
-const theme = createTheme(themeOverrides);
+
+const theme = createTheme(adaptV4Theme(themeOverrides));
 
 function App() {
     return (
-        <MuiThemeProvider theme={theme}>
-            <Router>
-                <Layout>
-                    <Route path="/arma-character-tool/" exact component={GuildPicker} />
-                    <Route path="/arma-character-tool/class-picker/" component={GuildPicker} />
-                    <Route path="/arma-character-tool/class-compare/" component={GuildCompare} />
-                    <Route path="/arma-character-tool/skill-picker/" component={SkillPicker} />
-                </Layout>
-            </Router>
-        </MuiThemeProvider>
+        <StyledEngineProvider injectFirst>
+            <ThemeProvider theme={theme}>
+                <Router>
+                    <Layout>
+                        <Route path="/" exact component={GuildPicker} />
+                        <Route path="/arma-character-tool/" exact component={GuildPicker} />
+                        <Route path="/arma-character-tool/class-picker/" component={GuildPicker} />
+                        <Route path="/arma-character-tool/class-compare/" component={GuildCompare} />
+                        <Route path="/arma-character-tool/skill-picker/" component={SkillPicker} />
+                    </Layout>
+                </Router>
+            </ThemeProvider>
+        </StyledEngineProvider>
     );
 }
 
