@@ -1,31 +1,31 @@
 const theme = {
-  "palette": {
-    "type": "dark",
-    "primary": {
-      "main": "#d84315",
-      "contrastText": "rgba(0, 0, 0, 0.87)"
+    "palette": {
+        "type": "dark",
+        "primary": {
+            "main": "#d84315",
+            "contrastText": "rgba(0, 0, 0, 0.87)"
+        },
+        "secondary": {
+            "main": "#5d4037"
+        },
+        "primary1Color": "#d84315",
+        "primary2Color": "#5d4037",
+        "accent1Color": "#eeeeee",
+        "accent2Color": "rgba(255, 236, 179, 0.3)",
+        "accent3Color": "#ff8f00"
     },
-    "secondary": {
-      "main": "#5d4037"
+    "tabs": {
+        "backgroundColor": "#e64a19",
+        "textColor": "rgba(33, 33, 33, 0.54)"
     },
-    "primary1Color": "#d84315",
-    "primary2Color": "#5d4037",
-    "accent1Color": "#eeeeee",
-    "accent2Color": "rgba(255, 236, 179, 0.3)",
-    "accent3Color": "#ff8f00"
-  },
-  "tabs": {
-    "backgroundColor": "#e64a19",
-    "textColor": "rgba(33, 33, 33, 0.54)"
-  },
-  "tableRow": {
-    "selectedColor": "rgba(255, 255, 255, 0.1)",
-    "hoverColor": "rgba(255, 236, 179, 0.15)"
-  },
-  "toggle": {
-    "thumbOffColor": "rgba(255, 236, 179, 0.51)",
-    "labelDisabledColor": "rgba(255, 255, 255, 0.29)",
-    "thumbDisabledColor": "rgba(255, 255, 255, 0.48)"
-  }
+    "tableRow": {
+        "selectedColor": "rgba(255, 255, 255, 0.1)",
+        "hoverColor": "rgba(255, 236, 179, 0.15)"
+    },
+    "toggle": {
+        "thumbOffColor": "rgba(255, 236, 179, 0.51)",
+        "labelDisabledColor": "rgba(255, 255, 255, 0.29)",
+        "thumbDisabledColor": "rgba(255, 255, 255, 0.48)"
+    }
 };
 export default theme;
